+++ CHANGELOG.md
@@ -5,6 +5,7 @@
 ### Added
 
 - [TD-6938] Add Grant Requests Search scroll and inserted_at filter
+- [TD-6780] Add Tag and Structure Tag Seach controller and context
 
 ## [6.15.0] 2024-11-27